feat(preset) add random mods button at charactors page
